python/requirements.txt
openpyxl==3.1.2
pillow==10.2.0
reportlab==4.0.9
